+++ src/fuzzy_finder/constructors/spawn_fzf.rs
@@ -1,7 +1,7 @@
+use crate::termtui::Parser;
 use anyhow::{Context, Result};
-use portable_pty::{native_pty_system, Child, CommandBuilder, MasterPty, PtySize};
+use portable_pty::{native_pty_system, CommandBuilder, PtySize};
 use std::sync::{Arc, Mutex};
-use vt100::Parser;
 
 use crate::fuzzy_finder::{FuzzyFinder, FuzzyFinderTerminal};
 
@@ -70,7 +70,7 @@ impl FuzzyFinder {
         writer.write_all(items_str.as_bytes())?;
         writer.flush()?;
 
-        let parser = Arc::new(Mutex::new(Parser::new(rows, cols, 0)));
+        let parser = Arc::new(Mutex::new(Parser::new(rows as usize, cols as usize, 0)));
 
         self.terminal = Some(FuzzyFinderTerminal {
             parser,
